Add auth check back

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\EditPostRequest;
 use App\Http\Requests\NewPostRequest;
 use App\Models\Post;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -17,6 +18,8 @@ class PostController extends Controller
 
     public function create()
     {
+        abort_unless(Auth::check(), 401);
+
         $post = new Post();
         return view('posts.create', compact('post'));
     }
@@ -27,6 +30,8 @@ class PostController extends Controller
      */
     public function store(NewPostRequest $request): RedirectResponse
     {
+        abort_unless(Auth::check(), 401);
+
         $post = new Post($request->except(['image', 'tags']));
         $request->user()->posts()->save($post);
 
@@ -49,6 +54,8 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+        abort_unless(Auth::check(), 401);
+
         return view('posts.edit', compact('post'));
     }
 
@@ -58,6 +65,8 @@ class PostController extends Controller
      */
     public function update(EditPostRequest $request, Post $post): RedirectResponse
     {
+        abort_unless(Auth::check(), 401);
+
         $post->update($request->except(['image', 'tags']));
 
         if($request->input('remove_old_image', false) && $post->hasMedia()) {
@@ -77,6 +86,8 @@ class PostController extends Controller
 
     public function destroy(Post $post): RedirectResponse
     {
+        abort_unless(Auth::check(), 401);
+
         $post->delete();
         return redirect()->route('posts.index');
     }
